+++ src/views/AuthPage.vue
@@ -1,0 +1,9 @@
+<template>
+  <ion-page>
+    <ion-router-outlet></ion-router-outlet>
+  </ion-page>
+</template>
+
+<script lang="ts" setup>
+import { IonRouterOutlet, IonPage } from "@ionic/vue";
+</script>
