+++ src/views/JoinedBroadcastsPage.vue
@@ -9,12 +9,12 @@
 /**************************************************/
 /* IMPORTS */
 /**************************************************/
-import {IonPage} from "@ionic/vue";
+import { IonPage } from "@ionic/vue";
 import NoBroadcasts from "@/components/NoBroadcasts.vue";
 import BroadcastCard from "@/components/BroadcastCard.vue";
 import { onMounted } from "vue";
 import { useBroadcastStore } from "@/store/broadcast";
-import {useMessageStore} from "@/store/message"
+import { useMessageStore } from "@/store/message";
 import router from "@/router";
 
 /**************************************************/
@@ -26,10 +26,10 @@ const messageStore = useMessageStore();
 /**************************************************/
 /* COMPONENT FUNCTIONS */
 /**************************************************/
-async function onClickJoin(broadcast:any){
-  await broadcastStore.join(broadcast.id)
-  messageStore.bcastId = broadcast.id
-  router.push('/home/chat-page')
+async function onClickJoin(broadcast: any) {
+  await broadcastStore.join(broadcast.id);
+  messageStore.bcastId = broadcast.id;
+  router.push("/home/chat-page");
 }
 
 /**************************************************/
